refactor: repository llm json endpoint
