+++ src/agents/__init__.py
@@ -4,6 +4,12 @@ from .organization_enrichment import (
     enrich_organizations,
     enrich_organizations_from_dict,
 )
+from .repository import (
+    llm_request_repo_infos,
+)
+from .user import (
+    llm_request_user_infos,
+)
 from .user_enrichment import enrich_users, enrich_users_from_dict
 
 __all__ = [
@@ -11,4 +17,6 @@ __all__ = [
     "enrich_organizations_from_dict",
     "enrich_users",
     "enrich_users_from_dict",
+    "llm_request_repo_infos",
+    "llm_request_user_infos",
 ]
