+++ src/llm/__init__.py
@@ -1,11 +1,25 @@
-"""LLM processing and repository analysis."""
+"""
+LLM processing and repository analysis.
+"""
 
-from .genai_model import (
-    llm_request_repo_infos,
-    llm_request_user_infos,
+from .model_config import (
+    ENV_VAR_MAPPINGS,
+    MODEL_CONFIGS,
+    create_pydantic_ai_model,
+    get_model_parameters,
+    get_retry_delay,
+    load_model_config,
+    validate_config,
 )
 
 __all__ = [
-    "llm_request_repo_infos",
-    "llm_request_user_infos",
+    # Configuration constants
+    "MODEL_CONFIGS",
+    "ENV_VAR_MAPPINGS",
+    # Configuration functions
+    "load_model_config",
+    "create_pydantic_ai_model",
+    "get_model_parameters",
+    "validate_config",
+    "get_retry_delay",
 ]
